pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "argocd-helm-template"
version = "0.1.0"
description = "Extract chart info from application.yaml, download chart, and run helm template"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "pyyaml",
    "click>=8.0",
]

[project.scripts]
argocd-helm-template = "argocd_helm_template:cli"

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
]

[tool.setuptools]
packages = ["argocd_helm_template"]
